# contract/openapi-1.2.0.json
{
  "components": {
    "schemas": {
...
      "DrillthroughRequest": {
        "properties": {
          "token": {
            "minLength": 1,
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "title": "DrillthroughRequest",
        "type": "object"
      },
      "DrillthroughResponse": {
        "properties": {
          "answer_id": {
            "title": "Answer Id",
            "type": "string"
          },
          "approximate": {
            "default": false,
            "title": "Approximate",
            "type": "boolean"
          },
          "row_count": {
            "title": "Row Count",
            "type": "integer"
          },
          "rows": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Rows",
            "type": "array"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sql_used": {
            "title": "Sql Used",
            "type": "string"
          },
          "total_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Count"
          }
        },
        "required": [
          "answer_id",
          "session_id",
          "sql_used",
          "row_count"
        ],
        "title": "DrillthroughResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Allowlisted DMS\u2194engine contract surface (operationIds: ask, drillthrough, ledger.append, ledger.verify, submit, tool.registry). cortex_contract 1.2.0. Engine semver is independent (see CortexOS.__version__ / docs/RELEASING.md).",
    "title": "Cortex Engine Contract API",
    "version": "1.2.0",
    "x-cortex-contract-routes": [
      "ask",
      "drillthrough",
      "ledger.append",
      "ledger.verify",
      "submit",
      "tool.registry"
    ],
    "x-cortex-contract-version": "1.2.0",
    "x-cortex-ports": [
      8000,
      8010
    ]
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/contract/drillthrough": {
      "post": {
        "description": "Re-derive contributing rows under the same session manifest (T7).",
        "operationId": "drillthrough",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DrillthroughRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DrillthroughResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Contract Drillthrough",
        "tags": [
          "contract"
        ]
      }
    },
...
  }
}